Poupou/user_service/requirements.txt
fastapi
uvicorn
asyncpg
python-dotenv
python-jose[cryptography]
pydantic[email]